requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

pydantic==2.5.3
pydantic-settings==2.1.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

celery[redis]==5.3.6
redis==5.0.1

httpx==0.26.0
aiofiles==23.2.1
aioboto3==12.3.0
python-multipart==0.0.6

openai==1.10.0
midtransclient==1.3.1
firebase-admin==6.3.0

structlog==24.1.0
sentry-sdk[fastapi]==1.39.2

pillow==10.2.0
python-magic==0.4.27

pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
